requirements.txt
matplotlib==3.8.4
altair==4.2.2
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
entrypoints==0.4
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
imagecodecs==2025.3.30
imageio==2.37.0
importlib_metadata==8.7.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
lazy_loader==0.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
narwhals==1.40.0
networkx==3.4.2
numpy==1.26.4
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==10.3.0
protobuf==3.20.3
pyarrow==20.0.0
pydeck==0.9.1
Pygments==2.19.1
Pympler==1.1
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.36.2
requests==2.32.3
rich==13.9.4
rpds-py==0.25.0
scikit-image==0.25.2
scipy==1.15.3
six==1.17.0
smmap==5.0.2
st-pages==1.0.1
streamlit==1.40.0
streamlit-camera-input-live==0.2.0
streamlit-drawable-canvas==0.9.3
streamlit-image-comparison==0.0.4
tenacity==8.5.0
tifffile==2025.5.10
toml==0.10.2
toolz==1.0.0
tornado==6.5
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
validators==0.35.0
zipp==3.21.0

fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
pydantic==2.7.1
httpx==0.27.0
aiofiles==23.2.1
shapely==2.1.1